+++ tests/unit/helpers/utils.py
@@ -511,7 +511,13 @@ def create_test_type_map(specs, container_classes, mappers=None):
     return type_map
 
 
-def create_load_namespace_yaml(namespace_name, specs, output_dir, incl_types, type_map):
+def create_load_namespace_yaml(
+    namespace_name: str,
+    specs: list[GroupSpec | DatasetSpec],
+    output_dir: str,
+    incl_types: dict[str, list[str] | None],
+    type_map: TypeMap,
+) -> dict:
     """
     Create a TypeMap with the specs loaded from YAML files and dependencies resolved.
 
@@ -521,6 +527,7 @@ def create_load_namespace_yaml(namespace_name, specs, output_dir, incl_types, ty
 
     :param namespace_name: Name of the new namespace.
     :param specs: List of specs of new data types to add.
+    :param output_dir: Directory to write the namespace and spec YAML files to.
     :param incl_types: Dict mapping included namespace name to list of data types to include or None to include all.
     :param type_map: The type map to load the namespace into.
     """
@@ -544,7 +551,7 @@ def create_load_namespace_yaml(namespace_name, specs, output_dir, incl_types, ty
 
     ns_builder.export(ns_filename, outdir=output_dir)
     ns_path = os.path.join(output_dir, ns_filename)
-    type_map.load_namespaces(ns_path)
+    return type_map.load_namespaces(ns_path)
 
 
 # ##### custom spec classes #####