+++ tests/unit/test_io_hdf5_h5tools.py
@@ -2216,7 +2216,10 @@ class TestLoadNamespaces(TestCase):
         """Test that loading namespaces given a path is OK and returns the correct dictionary."""
         ns_catalog = NamespaceCatalog()
         d = HDF5IO.load_namespaces(ns_catalog, self.path)
-        self.assertEqual(d, {'test_core': {}})  # test_core has no dependencies
+        # test_core has no dependencies
+        self.assertEqual(d, {'test_core': {}})
+        # source types should be stored on the catalog
+        self.assertTupleEqual(ns_catalog.get_source_types('test_core'), ('Foo', 'FooBucket', 'FooFile'))
 
     def test_load_namespaces_no_path_no_file(self):
         """Test that loading namespaces without a path or file raises an error."""
@@ -2264,7 +2267,7 @@ class TestLoadNamespaces(TestCase):
         pathlib_path = Path(self.path)
         ns_catalog = NamespaceCatalog()
         d = HDF5IO.load_namespaces(ns_catalog, pathlib_path)
-        self.assertEqual(d, {'test_core': {}})  # test_core has no dependencies
+        self.assertEqual(d, {'test_core': {}})
 
     def test_load_namespaces_with_dependencies(self):
         """Test loading namespaces where one includes another."""
@@ -2294,7 +2297,11 @@ class TestLoadNamespaces(TestCase):
 
         ns_catalog = NamespaceCatalog()
         d = HDF5IO.load_namespaces(ns_catalog, self.path)
-        self.assertEqual(d, {'test_core': {}, 'test_core2': {'test_core': ('Foo', 'FooBucket', 'FooFile')}})
+        self.assertEqual(d, {
+            'test_core': {},
+            'test_core2': {'test_core': ('Foo', 'FooBucket', 'FooFile')},
+        })
+        self.assertTupleEqual(ns_catalog.get_source_types('test_core'), ('Foo', 'FooBucket', 'FooFile'))
 
     def test_load_namespaces_no_specloc(self):
         """Test loading namespaces where the file does not contain a SPEC_LOC_ATTR."""
@@ -2340,12 +2347,16 @@ class TestLoadNamespaces(TestCase):
 
         # load the namespace from file
         ns_catalog = NamespaceCatalog(CustomGroupSpec, CustomDatasetSpec, CustomSpecNamespace)
-        namespace_deps = HDF5IO.load_namespaces(ns_catalog, self.path)
+        namespace_types = HDF5IO.load_namespaces(ns_catalog, self.path)
+
+        # test that the source types are correct
+        expected = ('Foo', 'FooBucket', 'FooFile', 'BigFoo', 'BiggerFoo')
+        self.assertTupleEqual(ns_catalog.get_source_types('test_core'), expected)
+        self.assertTupleEqual(ns_catalog.get_source_types('test-ext'), ('FooExt',))
 
         # test that the dependencies are correct
         expected = ('Foo',)
-        self.assertTupleEqual((namespace_deps['test-ext']['test_core']), expected)
-
+        self.assertTupleEqual(namespace_types['test-ext']['test_core'], expected)
         # test that the types are loaded
         types = ns_catalog.get_types('test-ext.extensions')
         expected = ('FooExt',)